# Speed QKV set_lora_info by eliminating unnecessary cudaStreamSyncrhonize 